requirements.txt
# Dépendances pour l'API Vercel (api/generate.py)
# Pour l'app Streamlit en local, utilisez: pip install -r requirements-local.txt
google-generativeai>=0.3.0
mistralai>=1.0.0
